Sergi T Commit 2 - GameOver Fadein Screen
Upon dying, the screen shifts to red and the words Game Over appear on the screen.

The speed of the fade-in is in the
variable called DeathDelay, located in the FirstPersonController.

The default falue is 0.1. The higher this value, the faster the fadein will hppen.
(a value of 1 is so fast that the player barely has time to see it.)

To do:

Add a sound effect along with the death.